changed i to size_t in declarations

+++ search_algorithms/1-binary.c
@@ -10,8 +10,8 @@
  */
 int binary_search(int *array, size_t size, int value)
 {
-	int left = 0, right = size - 1;
-	int mid;
+	size_t left = 0, right = size - 1;
+	size_t mid;
 	int i;
 
 	if (array == NULL)
